+++ server/woowa-orm/data-types.js
@@ -1,0 +1,9 @@
+module.exports = {
+  BOOLEAN: 'tinyint(1)',
+  INTEGER: 'int(11)',
+  DATE: 'date',
+  DATETIME: 'datetime',
+  TIMESTAMP: 'timestamp',
+  STRING: 'varchar(255)',
+  TEXT: 'text',
+}
